Added initial Store class member variables (floating point arrays with prices of items that decrease if you buy in bulk) and methods (for buying cups/ice/lemons/sugar)

+++ LemonadeStand/Inventory.cs
@@ -7,7 +7,7 @@ namespace LemonadeStand
 {
   public class Inventory
   {
-    private decimal money;
+    private float money;
     private ushort cups;
     private ushort lemons;
     private ushort sugar;
